refactor: Remove unused /docs/[docId]/evaluations page and components
- Delete entire evaluations page directory with all its components
- Move rerunEvaluation and createOrRerunEvaluation actions to new location
- Update import paths in EvaluationManagement and RerunButtonClient
- Fix MCP server API path from /api/docs to /api/documents for consistency
- Update README documentation to reflect correct API endpoint

The evaluations page was unused - evaluation functionality is already
available through the main docs page and individual eval pages.
