--- apps/web/src/app/docs/[docId]/evaluations/types.ts
@@ -1,10 +1,0 @@
-import type { Evaluation } from "@/shared/types/databaseTypes";
-
-export interface AgentWithEvaluation {
-  id: string;
-  name: string;
-  version: string;
-  description: string;
-  evaluation?: Evaluation;
-  isIntended: boolean;
-}
